requirements.txt
tkinter
sv_ttk
pillow
threading
selenium
pyinstaller